+++ R/db_info.R
@@ -103,7 +103,9 @@ print.rquery_db_info <- function(x, ...) {
   print(format(x))
 }
 
-
+#' An example \code{rquery_db_info} object useful for formatting \code{SQL} without a database connection.
+#'
+#' @export
 rquery_default_db_info <- rquery_db_info(identifier_quote_char = '"',
                                          string_quote_char = "'",
                                          is_dbi = FALSE)